issue#6対応
