+++ index.html
@@ -29,6 +29,6 @@
         </main>
     </div>
     <footer></footer>
-    <script src = "PitchCalc.js" type="module"></script>
+    <script src = "recording.js" type="module"></script>
 </body>
 </html>
